graph:
Added weight attribute to the iCelEdge class. The edge weight was being calculated during the ShortestPath() method, as the Euclidean distance between the two nodes.
Added another AddSuccessor() method to iCelNode, which takes the edge weigth as a parameter.
Added a GetEdges() method to iCelNode.
Changed iCelGraph::ShortestPath() to get the edge weight from iCelEdge instead of calculating it.
Added an AddEdge method to iCelGraph, which takes the edge weight as a parameter.

hpf:
Changes to iCelHNavStruct::BuildHighLevelGraph(). Still needs to be tested.