graph:
Added a ShortestPath2() method to iCelGraph. I was not sure the old one was working. Also, the original method return path doesn't include the origin of the path, while the path created with ShortestPath2 does.
Removed a few commented out methods and unused attributes.
Fixed iCelPath::Invert() (It was inverting the path twice, so it's output was the same as it's input).
Fixed iCelNode::Heuristic() (The cost attribute was not being updated correctly).

hpf:
Fixed iCelHNavStruct::BuildHighLevelGraph().
Fixed iCelHNavStruct::ShortestPath(). However, the iCelPath implementation returned by this method is not ready yet (at this moment, it only contains the high level path, indicating what sectors the path goes through).