Added the ability for entity templates to reuse other entity templates.
This adds the possibility of having master entity templates which are
used by various other entity templates for specialization. When an
entity template is merged with an existing entity template this merge
process makes sure that property classes with the same name (and tag) are
also merged (without the creation of a new property class).