The elcm registered entities for deletion which shouldn't be registered.
The idea is that it only registers entities for deletion which were
already present before the baseline. The problem is that the dynworld
plugin actually deletes (and creates too, but we already solved that
problem) entities for memory savings but these entities are not really
deleted. The current solution is a bit of a hack but it works.