Added (untested) support for picking up dynamic objects with modified entities.
In this situation the entity cannot be converted to a template (because it
has state) and so it needs to be put in the inventory as an entity itself.
Saving/loading is not yet implemented and also I didn't test it yet as
I have to create a testcase in elcmtest first.