swig: suppress warning about unhandled nested classes
SWIG can not yet handle nested classes gracefully, and warns when it
encounters them. Each SCF interface invokes SCF_INTERFACE() which
results in a nested InterfaceTraits class, hence SWIG emits a
significant number of such warnings. Since, in this case, these warnings
are just noise (that is, they are fundamental to SCF and will not be
removed), suppress the diagnostic. SWIG (annoyingly) also will warn about
other cases of nested classes, even if they have been %ignored, so this
suppresses that noise as well.